implement email sending on upload candidacy event

+++ src/Controller/CandidateController.php
@@ -2,11 +2,11 @@
 
 namespace App\Controller;
 
+use Exception;
 use App\Entity\Candidate;
 use App\Form\CandidateType;
 use App\Repository\CandidateRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -17,52 +17,57 @@ use Symfony\Component\HttpFoundation\File\Exception\FileException;
 class CandidateController extends AbstractController
 {
     #[Route('/candidate/{id}', name: 'app_candidate_{id}')]
-    public function index(CandidateRepository $repository, Request $request, Candidate $candidate, EntityManagerInterface $em): Response
-    {
-        
-        $id = $request->get('id');
-        $user = $repository->findOneBy(['id' => $id]);
+public function index(CandidateRepository $repository, Request $request, Candidate $candidate, EntityManagerInterface $em): Response
+{
+    $id = $request->get('id');
+    $user = $repository->findOneBy(['id' => $id]);
 
-        $form = $this->createForm(CandidateType::class, $candidate);
+    $form = $this->createForm(CandidateType::class, $candidate);
 
-        $form->handleRequest($request);
+    $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+    if ($form->isSubmitted() && $form->isValid()) {
 
-            $CVFile = $candidate->getCurriculumVitae();
+        $CVFile = $candidate->getCurriculumVitae();
 
-            if ($CVFile != null) {
-                $filePath = $this->getParameter('kernel.project_dir') . '/public/images/cv/' . $CVFile;
+        // Vérifier si un fichier existe déjà
+        if ($CVFile != null) {
+            $filePath = $this->getParameter('kernel.project_dir') . '/public/images/cv/' . $CVFile;
+            if (file_exists($filePath)) {
                 unlink($filePath);
             }
-
-            $uploadedFile = $form['curriculumVitaeFile']->getData();
-
-            if ($uploadedFile instanceof UploadedFile) {
-                $destination = $this->getParameter('kernel.project_dir') . '/public/images\cv';
-                $originalFilename = pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME);
-                $newFilename = $originalFilename.'-'.uniqid().'.'.$uploadedFile->guessExtension();
-
-                try {
-                    $uploadedFile->move($destination, $newFilename);
-                } catch (FileException $e) {
-                    throw New Exception('erreur lors de l\'envoi du fichier');
-                }
-
-                $candidate->setCurriculumVitae($newFilename);
-            }
-
-            $em->flush();
-            $this->addFlash('success', 'votre compte a bien été modifié.');
-            return $this->redirectToRoute('app_candidate_{id}', ['id' => $id]);
-
         }
 
-        return $this->render('candidate/index.html.twig', [
-            'user' => $user,
-            'candidateForm' => $form
-        ]);
+        $uploadedFile = $form['curriculumVitaeFile']->getData();
+
+        // Vérifier si un nouveau fichier a été téléchargé
+        if ($uploadedFile instanceof UploadedFile) {
+            // Générer un nouveau nom de fichier et enregistrer le fichier téléchargé
+            $destination = $this->getParameter('kernel.project_dir') . '/public/images/cv';
+            $originalFilename = pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME);
+            $newFilename = $originalFilename.'-'.uniqid().'.'.$uploadedFile->guessExtension();
+
+            try {
+                $uploadedFile->move($destination, $newFilename);
+            } catch (FileException $e) {
+                throw New Exception('erreur lors de l\'envoi du fichier');
+            }
+
+            $candidate->setCurriculumVitae($newFilename);
+        }
+
+        $em->flush();
+        $this->addFlash('success', 'votre compte a bien été modifié.');
+
+        return $this->redirectToRoute('app_candidate_{id}', ['id' => $id]);
     }
+
+    return $this->render('candidate/index.html.twig', [
+        'user' => $user,
+        'candidateForm' => $form
+    ]);
+}
+
 
 
     #[Route('/candidate', name: 'app_candidate')]
